Move editing support to seperate class + file
